skinedit, tiemu: fix a number of compiler (gcc & clang) warnings, memory leaks, etc. outside GDB + dependencies. Increase default warning level, use _FORTIFY_SOURCE=2 and stack protector.
git-svn-id: http://svn.tilp.info/repos/tiemu@2863 4f589446-7ac1-0310-82c7-a055b08196ef

+++ tiemu/trunk/src/core/ti_hw/m68k.c
@@ -205,7 +205,7 @@ int hw_m68k_run(int n, unsigned maxcycles)
             bkpts.id = 0;
             while(l)
             {
-                if(BKPT_ADDR(GPOINTER_TO_INT(l->data)) == (int)m68k_getpc())
+                if(BKPT_ADDR(GPOINTER_TO_INT(l->data)) == (unsigned int)m68k_getpc())
                 {
 					if(BKPT_IS_TMP(GPOINTER_TO_INT(l->data)))
 						bkpts.code = g_list_remove(bkpts.code, l->data);
@@ -228,7 +228,7 @@ int hw_m68k_run(int n, unsigned maxcycles)
 			uint16_t offset = GPOINTER_TO_INT(bkpts.pgmentry->data) & 0xffff;
 			bkpts.id = 0;
 
-			if(heap_deref(handle)+offset == (int)m68k_getpc())
+			if(heap_deref(handle)+offset == m68k_getpc()) // XXX CHECK from where handle and offset come from
 			{
 				bkpts.type = BK_TYPE_PGMENTRY;
 				return DBG_BREAK;
